added subimplementations for Scalar/Binary input

+++ Ohm/src/com/team1389/hardware/inputs/hardware/NavXHardware.java
@@ -19,7 +19,7 @@ public class NavXHardware implements CompositeWatchable {
 
 	public AngleIn getAngleInput() {
 		return new AngleIn(() -> {
-			return navX.getYaw();
+			return (double) navX.getYaw();
 		});
 	}
